+++ KLTN.Core/MissionServices/DTOs/MissionDetailResponseDTO.cs
@@ -22,7 +22,7 @@ namespace KLTN.Core.MissionServices.DTOs
         public int MaxStudentAmount { get; set; }
         public int JoinedStudentAmount { get; set; }
         public string LecturerName { get; set; }
-        public long TokenAmount { get; set; }
+        public decimal TokenAmount { get; set; }
         public List<JoinedStudentDTO> JoinedStudentList { get; set; }
         public bool IsJoined { get; set; }
     }
