+++ KLTN.Core/ScholarshipServices/DTOs/StudentScholarshipResponseDTO.cs
@@ -1,7 +1,5 @@
 ﻿using KLTN.DAL.Models.DTOs;
-using System;
 using System.Collections.Generic;
-using System.Text;
 
 namespace KLTN.Core.ScholarshipServices.DTOs
 {
@@ -21,7 +19,7 @@ namespace KLTN.Core.ScholarshipServices.DTOs
         public long EndTimeToComFirm { get; set; }
         public string LecturerInCharge { get; set; }
         public string LecturerName { get; set; }
-        public long TokenAmount { get; set; }
+        public decimal TokenAmount { get; set; }
         public int JoinedStudentAmount { get; set; }
         public List<JoinedStudentDTO> JoinedStudentList { get; set; }
         public bool IsJoined { get; set; }
