+++ KLTN.DAL/Models/Entities/Tuition.cs
@@ -1,9 +1,7 @@
 ﻿using KLTN.DAL.Models.DTOs;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
-using System;
 using System.Collections.Generic;
-using System.Text;
 
 namespace KLTN.DAL.Models.Entities
 {
@@ -23,8 +21,8 @@ namespace KLTN.DAL.Models.Entities
         public int SchoolYear { get; set; }
         public long StartTime { get; set; }
         public long EndTime { get; set; }
-        public long TokenAmount { get; set; }
-        public long CurrencyAmount { get; set; }
+        public decimal TokenAmount { get; set; }
+        public decimal CurrencyAmount { get; set; }
         public string LecturerInCharge { get; set; }
         public string LecturerName { get; set; }
         public int JoinedStudentAmount { get; set; }
